changing names, fixing styles
