--- client/app/components/AssuranceTemplate/styles.scss
@@ -1,5 +1,0 @@
-.aboutHero {
-  background-image: -webkit-linear-gradient(141deg, rgba(0, 0, 0, 0.3) 0%, rgba(0, 0, 0, 0.3) 71%, rgba(0, 0, 0, 0.3) 100%),-webkit-linear-gradient(rgba(0, 0, 0, 0.3), rgba(0, 0, 0, 0.3)), url('http://res.cloudinary.com/dyjo5a3ci/image/upload/v1514087320/idicia/about-us.png');
-  background-image: linear-gradient(141deg, rgba(0, 0, 0, 0.3) 0%, rgba(0, 0, 0, 0.3) 71%, rgba(0, 0, 0, 0.3) 100%),linear-gradient(rgba(0, 0, 0, 0.3), rgba(0, 0, 0, 0.3)), url('http://res.cloudinary.com/dyjo5a3ci/image/upload/v1514087320/idicia/about-us.png');
-  background-size: cover;
-}
