+++ client/app/components/SolutionsTemplate/SolutionCard/index.js
