+++ client/app/components/SolutionsTemplate/SolutionCard/styles.scss
