+++ client/app/components/SolutionsTemplate/styles.scss
@@ -1,61 +1,5 @@
-// Custom transforms for content blocks such as centered text and uppercased main headers
-.content {
-  > h1, h2, p {
-    text-align: center;
-  }
-
-  > h1 {
-    text-transform: uppercase;
-    font-weight: bold;
-  }
-
-  > p > a {
-    transition: all .3s ease-in-out;
-    background-color: transparent;
-    color: #3f3f3f;
-    border: 2px solid transparent;
-    border: 2px solid #3f3f3f;
-    padding: 14px 31px;
-    font-size: 13px;
-    font-weight: bold;
-    word-wrap: break-word;
-
-    &:hover {
-      background-color: #3f3f3f;
-      color: #ffffff;
-      border: 2px solid #3f3f3f;
-    }
-  }
-}
-
-.mainTitle {
-  text-align: center;
-  text-transform: uppercase;
-  font-weight: bold;
-}
-
-// Custom styling so that the six elements will render into two columns on larger screens only
-.columnCluster {
-  display: flex;
-  flex-flow: row wrap;
-
-  div {
-    display: inline-block;
-    width: 100%;
-    padding: 25px 10px;
-    @media (min-width: 769px) {
-	    width: 50%;
-	  }
-
-    > h2 > a {
-      color: #000066 !important;
-    }
-  }
-}
-
-.verticalCenter {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  height: 100%;
+.aboutHero {
+  background-image: -webkit-linear-gradient(141deg, rgba(0, 0, 0, 0.3) 0%, rgba(0, 0, 0, 0.3) 71%, rgba(0, 0, 0, 0.3) 100%),-webkit-linear-gradient(rgba(0, 0, 0, 0.3), rgba(0, 0, 0, 0.3)), url('http://res.cloudinary.com/dyjo5a3ci/image/upload/v1514087320/idicia/about-us.png');
+  background-image: linear-gradient(141deg, rgba(0, 0, 0, 0.3) 0%, rgba(0, 0, 0, 0.3) 71%, rgba(0, 0, 0, 0.3) 100%),linear-gradient(rgba(0, 0, 0, 0.3), rgba(0, 0, 0, 0.3)), url('http://res.cloudinary.com/dyjo5a3ci/image/upload/v1514087320/idicia/about-us.png');
+  background-size: cover;
 }
